Move map view to home view and share same view controller
