+++ NearMe/View Controller/HomeViewController.swift
@@ -10,25 +10,41 @@ import UIKit
 import GoogleMaps
 import GooglePlaces
 
+enum ViewType {
+    case List
+    case Map
+}
+
 class HomeViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, UISearchResultsUpdating {
 
     @IBOutlet weak var tableView: UITableView!
-    var searchController: UISearchController!
+    @IBOutlet weak var mapView: UIView!
+    
+    var currentViewType = ViewType.List
+    
     var posts = [Post]()
     var locationManager = CLLocationManager()
     var currentLocation: CLLocation?
     var currentRadius = 5.00
     let defaultLocation = CLLocation(latitude:37.3743507,longitude:-121.8825989)
+    
+    var mapContentView: GMSMapView!
+    var placesClient: GMSPlacesClient!
+    var zoomLevel: Float = 15.0
+    let DEFAULT_LATITUDE = 37.3743507
+    let DEFAULT_LONGITUDE = -121.8825989
+    var postLocations: [CLLocationCoordinate2D] = []
+    
+    var searchController: UISearchController!
     let refreshControl = UIRefreshControl()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.estimatedRowHeight = 100
-        tableView.rowHeight = UITableViewAutomaticDimension
-        
+       
         initLocation()
+        initTableView()
+        initMapView()
+        
         initRefreshControl()
     }
 
@@ -37,6 +53,16 @@ class HomeViewController: UIViewController, UITableViewDelegate, UITableViewData
         // Dispose of any resources that can be recreated.
     }
     
+    
+    /** MARK: - Table View **/
+    func initTableView(){
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.estimatedRowHeight = 100
+        tableView.rowHeight = UITableViewAutomaticDimension
+    }
+    
+
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return posts.count
     }
@@ -91,36 +117,101 @@ class HomeViewController: UIViewController, UITableViewDelegate, UITableViewData
     }
 
     func getPost(location: CLLocation, radius: Double) -> Void {
-        print("getting posts")
-        print("location: \(location), radius: \(radius)")
         PostService.sharedInstance.search(center: location, radius: radius, success: { (posts: [Post]) in
-            print("self.posts.count: \(self.posts.count)")
-            print("posts.count before assigning value: \(posts.count)")
             self.posts = posts.reversed()
             self.tableView.reloadData()
+            self.showPostsInMapView()
         }, failure: { (error: Error) in
             self.showError(error: error)
         })
     }
 
-    func refreshPost(location: CLLocation, radius: Double) -> Void {
-        print("refreshing posts")
-        print("location: \(location), radius: \(radius)")
-        PostService.sharedInstance.search(center: location, radius: radius, success: { (posts: [Post]) in
-            self.posts = posts.reversed()
-            self.tableView.reloadData()
-            self.refreshControl.endRefreshing()
-            print("posts.count: \(posts.count)")
-        }, failure: { (error: Error) in
-            self.showError(error: error)
-            print("posts.count: error")
-        })
+
+    @IBAction func onSwitchView(_ sender: Any) {
+        switch currentViewType {
+            case ViewType.List:
+                self.view.bringSubview(toFront: mapView)
+                currentViewType = ViewType.Map
+                break
+            case ViewType.Map:
+                self.view.bringSubview(toFront: tableView)
+                currentViewType = ViewType.List
+                break
+            default:
+                ()
+        }
+    }
+    
+    /** MARK: - Map View **/
+    func initMapView(){
+        locationManager = CLLocationManager()
+        locationManager.desiredAccuracy = kCLLocationAccuracyBest
+        locationManager.requestAlwaysAuthorization()
+        locationManager.distanceFilter = 50
+        locationManager.startUpdatingLocation()
+        locationManager.delegate = self
+        
+        placesClient = GMSPlacesClient.shared()
+        
+        let camera = GMSCameraPosition.camera(withLatitude: DEFAULT_LATITUDE,
+                                              longitude: DEFAULT_LONGITUDE,
+                                              zoom: zoomLevel)
+        mapContentView = GMSMapView.map(withFrame: view.bounds, camera: camera)
+        mapContentView.delegate = self
+        mapContentView.settings.myLocationButton = true
+        mapContentView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        mapContentView.isMyLocationEnabled = true
+        // Add the map to the view, hide it until we've got a location update.
+        mapView.addSubview(mapContentView)
     }
     
     
-    // MARK: - Pull and Refresh
-    func initRefreshControl() {
+    func showPostsInMapView() {
+        postLocations.removeAll()
+        for post in posts {
+            let state_marker = GMSMarker()
+            let post_latitude = post.location?.latitude ?? DEFAULT_LATITUDE
+            let post_longitude = post.location?.longitude ?? DEFAULT_LONGITUDE
+            let post_location = checkIfMutlipleCoordinates(latitude: post_latitude, longitude: post_longitude)
+            postLocations.append(post_location)
+            
+            state_marker.position = CLLocationCoordinate2D(latitude: post_location.latitude, longitude: post_location.longitude)
+            state_marker.title = post.screen_name
+            state_marker.snippet = post.message
+            state_marker.map = mapContentView
+        }
+    }
+    
+    func checkIfMutlipleCoordinates(latitude : Double , longitude : Double) -> CLLocationCoordinate2D{
         
+        var lat = latitude
+        var lng = longitude
+        var finalPos = CLLocationCoordinate2D()
+        
+        let arrTemp = self.postLocations.filter {
+            
+            return (((latitude == $0.latitude) && (longitude == $0.longitude)))
+        }
+        
+        if arrTemp.count > 0 {
+            // Core Logic giving minor variation to similar lat long
+            
+            let variation = (randomDouble(min: 0.0, max: 2.0) - 0.5) / 1500
+            lat = lat + variation
+            lng = lng + variation
+            finalPos = checkIfMutlipleCoordinates(latitude: lat, longitude: lng)
+        } else {
+            finalPos = CLLocationCoordinate2D(latitude: CLLocationDegrees(latitude), longitude: CLLocationDegrees(longitude))
+        }
+        return  finalPos
+    }
+    
+    func randomDouble(min: Double, max:Double) -> Double {
+        return (Double(arc4random()) / 0xFFFFFFFF) * (max - min) + min
+    }
+    
+    /** MARK: - Pull and Refresh **/
+    func initRefreshControl() {
         refreshControl.addTarget(self, action: #selector(refreshControlAction(_:)), for: UIControlEvents.valueChanged)
         tableView.insertSubview(refreshControl, at: 0)
     }
@@ -130,27 +221,38 @@ class HomeViewController: UIViewController, UITableViewDelegate, UITableViewData
          refreshPost(location: self.currentLocation!, radius: currentRadius)
     }
     
+    
+    func refreshPost(location: CLLocation, radius: Double) -> Void {
+        PostService.sharedInstance.search(center: location, radius: radius, success: { (posts: [Post]) in
+            self.posts = posts.reversed()
+            self.tableView.reloadData()
+            self.refreshControl.endRefreshing()
+        }, failure: { (error: Error) in
+            self.showError(error: error)
+        })
+    }
+    
+    
+    /** MARK: - Error window **/
     func showError(error: Error) {
         let alert = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
         let okAction = UIAlertAction(title: "OK", style: .default, handler: nil)
         alert.addAction(okAction)
         self.present(alert, animated: true, completion: nil)
     }
-
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-        switch segue.identifier! {
-            case "homeListToMap":
-                let mapViewNavigationController = segue.destination as! UINavigationController
-                let mapVC = mapViewNavigationController.topViewController as! MapViewController
-                mapVC.posts = posts
-            default:
-                ()
-        }
-        
-    }
 }
+
+extension HomeViewController: GMSMapViewDelegate {
+    func mapView(_ mapView: GMSMapView, markerInfoWindow marker: GMSMarker) -> UIView? {
+        let infoWindow = MapInfoWindow(frame: CGRect(x:0,y:0,width:200,height:65))
+        infoWindow.screenName = marker.title
+        infoWindow.message = marker.snippet
+        
+        return infoWindow
+    }
+    
+}
+
 
 extension HomeViewController: CLLocationManagerDelegate {
     
@@ -158,8 +260,14 @@ extension HomeViewController: CLLocationManagerDelegate {
     func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
         if let location = locations.last {
             self.currentLocation = location
-            print("currentLocation: \(self.currentLocation)")
             getPost(location: self.currentLocation!, radius: currentRadius)
+            
+            let camera = GMSCameraPosition.camera(withLatitude:location.coordinate.latitude, longitude: location.coordinate.longitude, zoom: zoomLevel)
+            if mapView.isHidden {
+                mapContentView.camera = camera
+            } else {
+                mapContentView.animate(to: camera)
+            }
         }
         
     }
@@ -171,6 +279,7 @@ extension HomeViewController: CLLocationManagerDelegate {
             print("Location access was restricted.")
         case .denied:
             print("User denied access to location.")
+            mapView.isHidden = false
         case .notDetermined:
             print("Location status not determined.")
         case .authorizedAlways: fallthrough
